whatsapp-http-server/requirements.txt
fastapi
uvicorn
pydantic
boto3
python-dotenv 